requirements.txt
Django==3.2.8
djangorestframework==3.12.4
psycopg2-binary==2.9.1
PyPDF2==1.26.0
pdfminer.six==20201018
spacy==3.1.3
tensorflow==2.6.0
transformers==4.10.3
celery==5.1.2
pytest==6.2.5
pytest-django==4.4.0
docker==5.0.2
docker-compose==1.29.2
gunicorn==20.1.0
python-decouple==3.4
django-environ==0.4.5
django-cors-headers==3.10.0
django-filter==2.4.0
django-redis==5.0.0
redis==3.5.3
django-storages==1.11.1
boto3==1.18.33
botocore==1.21.33
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-health-check==3.16.4
django-prometheus==2.1.0
prometheus-client==0.11.0
django-extensions==3.1.3
django-admin-honeypot==1.1.0
django-axes==5.23.0
django-allauth==0.48.0
django-rest-auth==0.9.5
django-two-factor-auth==1.13.1
django-otp==1.1.3
django-user-agents==2.2.0
user-agents==2.2.0
ua-parser==0.10.0
